apps/gateway-api/src/app/app.service.spec.ts
import { Test, TestingModule } from '@nestjs/testing'
import { AppService } from './app.service'

describe('AppService', () => {
  let service: AppService

  beforeEach(async () => {
    const module: TestingModule = await Test.createTestingModule({
      providers: [AppService],
    }).compile()

    service = module.get<AppService>(AppService)
  })

  it('should be defined', () => {
    expect(service).toBeDefined()
  })

  it('should return status', () => {
    const status = service.getStatus()
    // Customize this based on the actual expected status
    expect(status).toBeDefined()
    // Add more assertions as needed
  })

  // Add more test cases as needed

  // Don't forget to clean up any resources, if necessary
  afterEach(() => {
    // Add cleanup logic here
  })
})
